Creating structure for adding more noise sampling methods.
